supabase-integration/requirements.txt
supabase>=2.0.0
postgrest>=0.13.0
google-generativeai>=0.3.0
numpy>=1.26.0
python-dotenv>=1.0.0
requests>=2.31.0
google-auth>=2.15.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0
